skip getline defition in linux

+++ src/bsd/getline.cc
@@ -2,7 +2,7 @@
 #include <cstdlib>
 #include <cstring>
 
-int getline(char **lineptr, size_t *size, FILE *fp) {
+size_t getline(char **lineptr, size_t *size, FILE *fp) {
     if (!*lineptr) *lineptr = (char*)malloc(*size + 1);
     if (!*lineptr) return -1;
